Reorganize structure of the code base on template: https://github.com/Spikeysanju/Wiggles.git
