+++ app/src/main/AndroidManifest.xml
@@ -14,12 +14,12 @@
         tools:targetApi="31" >
 
         <activity
-            android:name=".MainActivity2"
+            android:name=".view.HomePage"
             android:exported="false"
             android:label="@string/title_activity_main2"
             android:theme="@style/Theme.FarmConnect" />
         <activity
-            android:name=".SignIn"
+            android:name=".view.SignIn"
             android:exported="false"
             android:label="@string/title_activity_sign_in"
             android:theme="@style/Theme.FarmConnect" />
