+++ app/src/main/java/com/example/farmconnect/MainActivity.kt
@@ -21,6 +21,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.farmconnect.ui.theme.FarmConnectTheme
+import com.example.farmconnect.view.HomePage
+import com.example.farmconnect.view.SignIn
 import com.google.firebase.auth.FirebaseAuth
 
 
@@ -33,10 +35,7 @@ class MainActivity : ComponentActivity() {
         setContent {
             FarmConnectTheme {
                 // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
+                Surface(modifier = Modifier.fillMaxSize(), color = MaterialTheme.colorScheme.background) {
                     MainAct("Android")
                 }
             }
@@ -45,22 +44,19 @@ class MainActivity : ComponentActivity() {
         var isLoggedIn = mFirebaseAuth.currentUser
         Log.d(TAG, mFirebaseAuth.currentUser?.displayName.toString())
 
-        if (isLoggedIn == null) {
+       /* if (isLoggedIn == null) {
             // show sign in screen
             Log.d(TAG, "Sign in screen")
             startActivity(Intent(this, SignIn::class.java))
             finish()
 
-        } else {
+        } else {*/
             // show homepage screen
             Log.d(TAG, "HomePage screen")
-            startActivity(Intent(this, MainActivity2::class.java))
+            startActivity(Intent(this, HomePage::class.java))
             finish()
-        }
-
+        //}
     }
-
-
 }
 
 @Composable
