+++ app/src/main/java/com/example/farmconnect/model/MainViewModel.kt
@@ -1,8 +1,9 @@
-package com.example.farmconnect
+package com.example.farmconnect.model
 
 import androidx.annotation.DrawableRes
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.farmconnect.R
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.asStateFlow
