+++ app/src/main/java/com/example/farmconnect/view/HomePage.kt
@@ -1,26 +1,20 @@
-package com.example.farmconnect
+package com.example.farmconnect.view
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.annotation.StringRes
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.DrawerValue
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalNavigationDrawer
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.material3.rememberDrawerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
@@ -28,19 +22,20 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
-import com.example.farmconnect.ui.AppBar
-import com.example.farmconnect.ui.DrawerContent
+import com.example.farmconnect.nevigation.AppBar
+import com.example.farmconnect.nevigation.DrawerContent
 import com.example.farmconnect.ui.theme.FarmConnectTheme
 import androidx.navigation.compose.rememberNavController
-import com.example.farmconnect.ui.CharityModeScreen
-import com.example.farmconnect.ui.FarmModeScreen
-import com.example.farmconnect.ui.FinanceStatsScreen
-import com.example.farmconnect.ui.InventoryScreen
-import com.example.farmconnect.ui.MarketplaceScreen
+import com.example.farmconnect.R
+import com.example.farmconnect.ui.charity.CharityModeScreen
+import com.example.farmconnect.ui.farmer.FarmModeScreen
+import com.example.farmconnect.ui.farmer.FinanceStatsScreen
+import com.example.farmconnect.ui.farmer.InventoryScreen
+import com.example.farmconnect.ui.farmer.MarketplaceScreen
 import com.example.farmconnect.ui.SettingsScreen
 import kotlinx.coroutines.launch
 
-class MainActivity2 : ComponentActivity() {
+class HomePage : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
@@ -63,7 +58,7 @@ enum class Screens(@StringRes val title: Int) {
     Donate(title = R.string.Donate),
     //    Charity mode screens:
     Charity(title = R.string.charity),
-    Settings(title=R.string.settings)
+    Settings(title= R.string.settings)
 }
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -73,7 +68,7 @@ fun App() {
     val backStackEntry by navController.currentBackStackEntryAsState()
     // Get the name of the current screen
     val currentScreen = Screens.valueOf(
-        backStackEntry?.destination?.route ?:Screens.Farm.name
+        backStackEntry?.destination?.route ?: Screens.Farm.name
     )
     val drawerState = rememberDrawerState(DrawerValue.Closed)
     val scope = rememberCoroutineScope()
